# CompProgEdu/WebApp/src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CompProgEdu API",
    "version": "v1"
  },
  "paths": {
...
    "/api/courses/instructor/{instructorId}": {
      "get": {
        "tags": [
          "Course"
        ],
        "operationId": "GetAllByInstructorId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "instructorId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CourseSummaryDtoListValidateableResponse"
            }
          }
        }
      }
    },
...
    "/api/desired-outputs/{id}": {
      "get": {
        "tags": [
          "DesiredOutputs"
        ],
        "operationId": "GetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DesiredOutputGetDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/desired-outputs/by-assignment/{assignmentId}": {
      "get": {
        "tags": [
          "DesiredOutputs"
        ],
        "operationId": "GetByAssignmentId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "assignmentId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DesiredOutputGetDtoListValidateableResponse"
            }
          }
        }
      }
    },
    "/api/desired-outputs": {
      "post": {
        "tags": [
          "DesiredOutputs"
        ],
        "operationId": "Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateDesiredOutputRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DesiredOutputGetDtoValidateableResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "DesiredOutputs"
        ],
        "operationId": "Update",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateDesiredOutputRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DesiredOutputGetDtoValidateableResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DesiredOutputs"
        ],
        "operationId": "Delete",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DesiredOutputGetDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/desired-outputs/order": {
      "put": {
        "tags": [
          "DesiredOutputs"
        ],
        "operationId": "UpdateOrder",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DesiredOutputGetDto"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DesiredOutputGetDtoListValidateableResponse"
            }
          }
        }
      }
    },
    "/api/instructors": {
      "post": {
        "tags": [
          "Instructors"
        ],
        "operationId": "Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateInstructorRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/InstructorResponseDtoValidateableResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Instructors"
        ],
        "operationId": "GetAll",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/InstructorGetAllResponseDtoListValidateableResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Instructors"
        ],
        "operationId": "Delete",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/InstructorResponseDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/instructors/add-student-to-course": {
      "post": {
        "tags": [
          "Instructors"
        ],
        "operationId": "AddStudentToCourse",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddStudentToCourseRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentResponseDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/instructors/extract": {
      "post": {
        "tags": [
          "Instructors"
        ],
        "operationId": "ExtractFromFile",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "string",
            "format": "binary"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentDetailDtoListValidateableResponse"
            }
          }
        }
      }
    },
    "/api/instructors/upload": {
      "post": {
        "tags": [
          "Instructors"
        ],
        "operationId": "AddStudentsFromFile",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "StudentList",
            "type": "string"
          },
          {
            "in": "query",
            "name": "CourseId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentDetailDtoListValidateableResponse"
            }
          }
        }
      }
    },
    "/api/instructors/{id}": {
      "get": {
        "tags": [
          "Instructors"
        ],
        "operationId": "GetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/InstructorResponseDtoValidateableResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Instructors"
        ],
        "operationId": "Update",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/InstructorDetailDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/InstructorResponseDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/instructors/all-course-students/{courseId}": {
      "get": {
        "tags": [
          "Instructors"
        ],
        "operationId": "GetCourseStudents",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentResponseDtoListPaginatedResponse"
            }
          }
        }
      }
    },
...
    "/api/method-test-case/by-assignment/{assignmentId}": {
      "get": {
        "tags": [
          "MethodTestCase"
        ],
        "operationId": "GetByAssignmentId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "assignmentId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/MethodTestCaseGetDtoListValidateableResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "fieldName": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "CourseSummaryDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "section": {
          "type": "string"
        },
        "studentCount": {
          "format": "int32",
          "type": "integer"
        },
        "instructorName": {
          "type": "string"
        }
      }
    },
    "CourseSummaryDtoListValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CourseSummaryDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
...
    "DesiredOutputGetDto": {
      "required": [
        "input",
        "output"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentId": {
          "format": "int32",
          "type": "integer"
        },
        "input": {
          "type": "string"
        },
        "output": {
          "type": "string"
        },
        "pointValue": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "DesiredOutputGetDtoValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/DesiredOutputGetDto"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "DesiredOutputGetDtoListValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DesiredOutputGetDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "CreateDesiredOutputRequest": {
      "required": [
        "input",
        "output"
      ],
      "type": "object",
      "properties": {
        "assignmentId": {
          "format": "int32",
          "type": "integer"
        },
        "input": {
          "type": "string"
        },
        "output": {
          "type": "string"
        },
        "pointValue": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UpdateDesiredOutputRequest": {
      "required": [
        "input",
        "output"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentId": {
          "format": "int32",
          "type": "integer"
        },
        "input": {
          "type": "string"
        },
        "output": {
          "type": "string"
        },
        "pointValue": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CreateInstructorRequest": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "InstructorResponseDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "InstructorResponseDtoValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/InstructorResponseDto"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "InstructorGetAllResponseDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "InstructorGetAllResponseDtoListValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InstructorGetAllResponseDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "AddStudentToCourseRequest": {
      "type": "object",
      "properties": {
        "courseId": {
          "format": "int32",
          "type": "integer"
        },
        "emailAddress": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "studentSchoolNumber": {
          "type": "string"
        },
        "changedPassword": {
          "type": "boolean"
        }
      }
    },
    "StudentResponseDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "emailAddress": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "studentSchoolNumber": {
          "type": "string"
        },
        "changedPassword": {
          "type": "boolean"
        }
      }
    },
    "StudentResponseDtoValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/StudentResponseDto"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "StudentDetailDto": {
      "type": "object",
      "properties": {
        "emailAddress": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "studentSchoolNumber": {
          "type": "string"
        },
        "changedPassword": {
          "type": "boolean"
        }
      }
    },
    "StudentDetailDtoListValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StudentDetailDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "InstructorDetailDto": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "StudentResponseDtoListPaginatedResponse": {
      "type": "object",
      "properties": {
        "totalCount": {
          "format": "int32",
          "type": "integer"
        },
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StudentResponseDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
...
    "MethodTestCaseGetDto": {
      "required": [
        "output",
        "paramInputs"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "returnType": {
          "type": "string"
        },
        "methodTestInjectable": {
          "type": "string"
        },
        "assignmentId": {
          "format": "int32",
          "type": "integer"
        },
        "paramInputs": {
          "type": "string"
        },
        "output": {
          "type": "string"
        },
        "hint": {
          "type": "string"
        },
        "pointValue": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "MethodTestCaseGetDtoListValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MethodTestCaseGetDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer {token}"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}